[Refactor] [FE-35] 프로그램 코스 코드 분리
