+++ src/types/program/ProgramCost.ts
@@ -1,0 +1,1 @@
+export type ProgramCostType = 'GUIDE' | 'ALL';
